🎉 Release 6.0.1 (#238)
This PR was opened by the [ready-release-go](https://github.com/woodpecker-ci/plugin-ready-release-go) plugin. When you're ready to do a release, you can merge this pull-request and a new release with version `6.0.1` will be created automatically. If you're not ready to do a release yet, that's fine, whenever you add more changes to `main` this pull-request will be updated.

## Options

- [ ] Mark this version as a release candidate

## [6.0.1](https://codeberg.org/woodpecker-plugins/docker-buildx/releases/tag/v6.0.1) - 2025-05-04

### Misc

- Fix build arg flags buils [[#237](https://codeberg.org/woodpecker-plugins/docker-buildx/pulls/237)]

Reviewed-on: https://codeberg.org/woodpecker-plugins/docker-buildx/pulls/238
Co-authored-by: oauth <woodpecker-bot@obermui.de>
Co-committed-by: oauth <woodpecker-bot@obermui.de>